working on enemies, fixed a bug where the player x pos being determined by window width caused a desync between it and ground, and it couldent stop itself before contacting an object when moving
b/c it wasnt a multiple of ten, and the ground blocks are 100px accross, and player moves 10px every tick. I might add a systen to determin what the error would be on movement and correct for it to stop just before the block, but that'll take a while and add a lot of processing time. Not that the game is hard to run or anything tho lmao, this thing b running great on my shitty ass laptop :D